+++ cmd/atb/main.go
@@ -17,7 +17,7 @@ import (
 )
 
 const (
-	version          = "1.1.0"
+	version          = "1.1.1"
 	verifyFormatText = "text"
 	verifyFormatJSON = "json"
 	verifyAlgorithm  = "SHA-256||RFC8785"